Moved MyReservationsView to reservation.py
Also removed `overview.py`, as it's now empty.
