+++ make_queue/urls.py
@@ -79,7 +79,7 @@ urlpatterns = [
     path('<Reservation:reservation>/change/', login_required(reservation.reservation.ChangeReservationView.as_view()), name="change_reservation"),
     path('<int:pk>/finish/', login_required(reservation.reservation.MarkReservationFinishedView.as_view()), name="mark_reservation_finished"),
     path('<int:pk>/', login_required(reservation.reservation.DeleteReservationView.as_view()), name="delete_reservation"),
-    path('me/', login_required(reservation.overview.MyReservationsView.as_view()), name="my_reservations"),
+    path('me/', login_required(reservation.reservation.MyReservationsView.as_view()), name="my_reservations"),
     path('admin/',
          permission_required('make_queue.can_create_event_reservation', raise_exception=True)(admin.reservation.AdminReservationView.as_view()),
          name="admin_reservation"),
