+++ make_queue/views/reservation/__init__.py
@@ -1,4 +1,4 @@
-__all__ = ["calendar", "machine", "overview", "reservation", "rules"]
+__all__ = ["calendar", "machine", "reservation", "rules"]
 
 for _import in __all__:
     __import__(f"{__package__}.{_import}")
